Create Insertion Sort Algorithm in Data Structures using Python
Sorting and Searching algorithms like insertion sort, selection sort, merge sort are very important to learn as a part of the preparation of your placements and also for internships because these form the base of any programmer. As a programmer, you have to deal with large amounts of data. Hence, we use sorting algorithms and one of them is an insertion sort. Now, the Insertion sort algorithm is a comparison-based sorting algorithm in which the sorted array is built by inserting one element at its correct position at a time.